[#76] Moving Daefen to src/vendors and cleaning up contract tests
