+++ universal-login-example/.eslintrc.json
@@ -44,4 +44,5 @@
 
         }
     }
+
 }
