+++ universal-login-example/src/vendors/Daefen.js
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 /*
 Daefen
 ------
